feat: add K-Nearest Neighbors classifier and regressor
- Add KNNClassifier and KNNRegressor with configurable k, distance metric, and weighting strategy
- Add aggregators: classProportions, weightedClassProportions, weightedAvgPreds
- Add KNN model types, default settings, UI components (settings, metrics, plots)
- Add unit tests for KNNClassifier, KNNRegressor, and all aggregators

+++ src/ml/models/k-means/KMeans.ts
@@ -145,7 +145,7 @@ export class KMeans implements Model<Tensor2D> {
     private computeCentroids(data: Tensor2D, assignments: Tensor2D): Tensor2D {
         return tidy(() => {
             const labels = assignments.squeeze();
-            const hot = oneHot(labels, this.numClusters);
+            const hot = oneHot(labels, this.numClusters).toFloat();
             const counts = hot.sum(0).expandDims(1);
             const summed = hot.transpose().matMul(data);
